Modified add_comm method to work with new qasync
New version of qasync library (>=0.19) runs the asyncio loop not when
its QEventLoop is created, but when the created QEventLoop object
run_forever method is called. This causes problems, as ts_salobj needs
running loop when creating remote connection. This change delays remote
construction to time when the loop is running.